requirements.txt
amqp==5.1.1
asgiref==3.7.2
b3api==0.0.4
billiard==4.1.0
black==23.7.0
celery==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.4
django-celery-results==2.5.1
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
isort==5.12.0
kombu==5.3.2
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pytest==7.4.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytest-mock
redis>=3.5,<4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
